4 commit, Se agregaron pruebas y se mejoraron especto de la salida de consola, asi mismo se agregaron varios metodos nuevos al la clase AutoDao.

+++ src/main/java/com/AutosQA/db/CrearTabla.java
@@ -7,10 +7,10 @@ public class CrearTabla {
     public static void crearTabla() {
         String sql = """
             CREATE TABLE IF NOT EXISTS auto (
-                id BIGINT PRIMARY KEY AUTO_INCREMENT,
+                id INTEGER PRIMARY KEY AUTO_INCREMENT,
                 marca VARCHAR(100),
                 modelo VARCHAR(100),
-                fabricacion BIGINT
+                fabricacion INTEGER
             );
         """;
